Add pausing
The game will handle X11 events while paused, but it doesn't handle any
input but unpausing. The game physics will never run while the game is
paused because it runs after getting the next event fails. Because the
game is waiting for new events in paused state, getting the event won't
fail in normal circumstances.