Implement borderless windows
Toggling override-redirect with ChangeWindowAttributes doesn't seem to
work, so the game uses six windows, and only three of the windows are
mapped at a time. Unmapped windows are only moved and resized when they
are mapped.

Currently enabling window borders while the game is paused causes the
windows to teleport to the center on GNOME. I'm not sure what's the best
way to fix it.